Move events

+++ lib/core/src/core/struct/molecule.cpp
@@ -73,7 +73,7 @@ namespace VTX::Core::Struct
 
 	int Molecule::getUnknownResidueSymbolIndex( const std::string & p_symbol ) const
 	{
-		int residueIndex;
+		size_t residueIndex;
 
 		for ( residueIndex = 0; residueIndex < _unknownResidueSymbol.size(); residueIndex++ )
 		{
@@ -90,7 +90,7 @@ namespace VTX::Core::Struct
 
 	ChemDB::UnknownResidueData * const Molecule::getUnknownResidueSymbol( const std::string & p_symbol ) const
 	{
-		for ( int residueIndex = 0; residueIndex < _unknownResidueSymbol.size(); residueIndex++ )
+		for ( size_t residueIndex = 0; residueIndex < _unknownResidueSymbol.size(); residueIndex++ )
 		{
 			if ( _unknownResidueSymbol[ residueIndex ]->symbolStr == p_symbol )
 				return _unknownResidueSymbol[ residueIndex ];
@@ -101,7 +101,7 @@ namespace VTX::Core::Struct
 
 	int Molecule::addUnknownResidueSymbol( ChemDB::UnknownResidueData * const p_residueData )
 	{
-		int residueIndex;
+		size_t residueIndex;
 
 		for ( residueIndex = 0; residueIndex < _unknownResidueSymbol.size(); residueIndex++ )
 		{
@@ -311,7 +311,7 @@ namespace VTX::Core::Struct
 	{
 		int res = 0;
 
-		for ( int chainIdx = 0; chainIdx < _chains.size(); chainIdx++ )
+		for ( size_t chainIdx = 0; chainIdx < _chains.size(); chainIdx++ )
 		{
 			const Chain * const chain = _chains[ chainIdx ];
 			if ( chain == nullptr )
@@ -325,7 +325,7 @@ namespace VTX::Core::Struct
 	{
 		int res = 0;
 
-		for ( int residueIdx = 0; residueIdx < _residues.size(); residueIdx++ )
+		for ( size_t residueIdx = 0; residueIdx < _residues.size(); residueIdx++ )
 		{
 			const Residue * const residue = _residues[ residueIdx ];
 			if ( residue == nullptr )
